pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-notes"
version = "0.1.0"
description = "A CLI tool for transcribing voice recordings using Whisper and related technologies"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Mike Muryn", email = "mike.muryn@example.com"}
]
maintainers = [
    {name = "Mike Muryn", email = "mike.muryn@example.com"}
]
keywords = ["whisper", "transcription", "audio", "speech-to-text", "diarization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"
dependencies = [
    "openai-whisper==20231117",
    "whisperx",
    "faster-whisper",
    "torch",
    "torchaudio",
    "pyannote.audio",
    "rich",
    "pydantic>=2",
    "python-dotenv",
    "openai",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "safety>=2.0.0",
    "pydocstyle>=6.3.0",
    "radon>=6.0.0",
     "pip-audit>=2.6.0",
]

[project.scripts]
voice-notes = "voice_notes.wrapper:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["voice_notes"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=src/voice_notes",
    "--cov-fail-under=95",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "audio: marks tests as audio processing tests",
]
